Bug fixes and minor updates. Working fine

+++ extractTopBestIterates.m
@@ -1,15 +1,35 @@
 function extractTopBestIterates(varargin)
-
-defaults = {'nTopIteratesToCopy',10
+%Updated to use for overall or for generation
+defaults = {'lookAtGeneration', []                         %Leave empty for overall, or pass in the integer value of the generation
+            'nTopIteratesToCopy',10
             'folderToCopyBestIteratesTo','bestIterates'
-            'finishedRunningDirectory','finishedRunning'};
+            'finishedRunningDirectory','finishedRunning'
+            'removeOldBestIterateFolder',false,
+            'gaFilename','gaInformation.mat'};
 setDefaultsForVarargin(defaults,varargin);        
-[~,~,~,~,~,~,~,allIterateInfo] = processGaInformationForMatlabGaV2('suppressPlot',true,'filename','gaInformation.mat');
+[~,~,~,~,~,~,~,allIterateInfo] = processGaInformationForMatlabGaV2(...
+    'suppressPlot',true,...
+    'filename',gaFilename,...
+    'processGeneration',lookAtGeneration);
+load(gaFilename)
+%If we want to remove all the old best iterates and check if we need to
+%start a folder
+if removeOldBestIterateFolder
+    systemCommand = sprintf('rm -r %s',folderToCopyBestIteratesTo);
+    system(systemCommand);
+end
 
+if ~exist(folderToCopyBestIteratesTo,'dir');
+    mkdir(folderToCopyBestIteratesTo);
+end
 
-systemCommand = sprintf('rm -r %s',folderToCopyBestIteratesTo);
-system(systemCommand);
-mkdir(folderToCopyBestIteratesTo);
+%Double check that nTopIteratesToCopy is less than or equal to the number
+%of iterates avaiable 
+if nTopIteratesToCopy > gaInfo.daq.header.populationSize; 
+    nTopIteratesToCopy = gaInfo.daq.header.populationSize;
+end
+
+%Loop through iterates and copy
 for iIter = 1:nTopIteratesToCopy
     %Construct foldername
     for iScore = 1:length(allIterateInfo{iIter}.iterate)
@@ -17,9 +37,14 @@ for iIter = 1:nTopIteratesToCopy
         gen  = allIterateInfo{iIter}.generation(iScore);
         folderName = sprintf('GA_gen_%03i_iter_%03i',gen,iter);
         scoreString = strrep(num2str(allIterateInfo{iIter}.score(iScore)),'.',',');
-        newName = sprintf('RANK_%03i_%s_Score_%s',iIter,folderName,scoreString);
+        if isempty(lookAtGeneration) 
+            newName = sprintf('OVERALL_RANK_%03i_%s_Score_%s',iIter,folderName,scoreString);
+        else
+            newName = sprintf('GA_gen_%03i_RANK_%03i_iter_%03i_Score_%s',gen,iIter,iter,scoreString);
+        end
+        
         systemCommand = sprintf('cp -r ./%s/%s ./%s/%s',finishedRunningDirectory,folderName,folderToCopyBestIteratesTo,newName);
-        system(systemCommand)
+        system(systemCommand);
     end
     
 end
